add general outlier detectors and helm charts
