+++ components/outlier-detection/isolation-forest/OutlierIsolationForest.py
@@ -1,28 +1,23 @@
 import numpy as np
-import pickle
-from sklearn.ensemble import IsolationForest
 
+from CoreIsolationForest import CoreIsolationForest
 from utils import flatten, performance, outlier_stats
 
-class OutlierIsolationForest(object):
+class OutlierIsolationForest(CoreIsolationForest):
     """ Outlier detection using Isolation Forests.
     
-    Arguments:
-        - threshold (float): anomaly score threshold; scores below threshold are outliers
+    Parameters
+    ----------
+        threshold (float) : anomaly score threshold; scores below threshold are outliers
      
-    Functions:
-        - predict: detect and return outliers
-        - send_feedback: add target labels as part of the feedback loop
-        - metrics: return custom metrics
+    Functions
+    ----------
+        send_feedback : add target labels as part of the feedback loop
+        metrics : return custom metrics
     """
-    def __init__(self,threshold=0.,load_path='./models/'):
+    def __init__(self,threshold=0.,model_name='if',load_path='./models/'):
         
-        self.threshold = threshold
-        self.N = 0 # total sample count up until now
-        
-        # load pre-trained model
-        with open(load_path + 'model.pickle', 'rb') as f:
-            self.clf = pickle.load(f)
+        super().__init__(threshold=threshold, model_name=model_name, load_path=load_path)
         
         self._predictions = []
         self._labels = []
@@ -30,39 +25,31 @@ class OutlierIsolationForest(object):
         self.roll_window = 100
         self.metric = [float('nan') for i in range(18)]
     
-    def predict(self,X,feature_names):
-        """ Detect outliers from mse using the threshold. 
-        
-        Arguments:
-            - X: input data
-            - feature_names
-        """
-        self.decision_val = self.clf.decision_function(X) # anomaly scores
-        self._anomaly_score.append(self.decision_val)
-        self._anomaly_score = flatten(self._anomaly_score)
-        
-        # make prediction
-        self.prediction = (self.decision_val < self.threshold).astype(int) # scores below threshold are outliers
-        self._predictions.append(self.prediction)
-        self._predictions = flatten(self._predictions)
-        
-        self.N+=self.prediction.shape[0] # update counter
-        
-        return self.prediction
     
     def send_feedback(self,X,feature_names,reward,truth):
         """ Return outlier labels as part of the feedback loop.
         
-        Arguments:
-            - X: input data
-            - feature_names
-            - reward
-            - truth: outlier labels
+        Parameters
+        ----------
+            X : array of the features sent in the original predict request
+            feature_names : array of feature names. May be None if not available.
+            reward (float): the reward
+            truth : array with correct value (optional)
         """
+        _ = super().send_feedback(X,feature_names,reward,truth)
+        
+        # historical reconstruction errors and predictions
+        self._anomaly_score.append(self.decision_val)
+        self._anomaly_score = flatten(self._anomaly_score)
+        self._predictions.append(self.prediction)
+        self._predictions = flatten(self._predictions)
+        
+        # target labels
         self.label = truth
         self._labels.append(self.label)
         self._labels = flatten(self._labels)
         
+        # performance metrics
         scores = performance(self._labels,self._predictions,roll_window=self.roll_window)
         stats = outlier_stats(self._labels,self._predictions,roll_window=self.roll_window)
         
@@ -71,8 +58,9 @@ class OutlierIsolationForest(object):
         for c in convert: # convert from np to native python type to jsonify
             metric.append(np.asscalar(np.asarray(c)))
         self.metric = metric
-            
-        return
+        
+        return []
+    
     
     def metrics(self):
         """ Return custom metrics.
@@ -87,8 +75,8 @@ class OutlierIsolationForest(object):
             dec_val = float('nan')
             y_true = float('nan')
         else:
-            pred = int(self._predictions[-2])
-            dec_val = self._anomaly_score[-2]
+            pred = int(self._predictions[-1])
+            dec_val = self._anomaly_score[-1]
             y_true = int(self.label[0])
                          
         is_outlier = {"type":"GAUGE","key":"is_outlier","value":pred}
