+++ components/outlier-detection/isolation-forest/requirements.txt
@@ -1,6 +1,6 @@
 numpy==1.14.5
 argparse==1.1
 pandas==0.23.4
-scikit-learn==0.19.1
+scikit-learn==0.20.1
 scipy==1.1.0
 requests>=2.20.0
