+++ components/outlier-detection/isolation-forest/train.py
@@ -13,13 +13,6 @@ random.seed(2018)
 # default args
 DATASET = 'kddcup99'
 SAMPLES = 50000
-SAVE = True
-SAVE_PATH = './models/'
-# Isolation Forest hyperparameters
-CONTAMINATION = .1
-N_ESTIMATORS = 50
-MAX_SAMPLES = .8
-MAX_FEATURES = 1.
 COLS = str(['duration','protocol_type','flag','src_bytes','dst_bytes','land','wrong_fragment','urgent','hot',
     'num_failed_logins','logged_in','num_compromised','root_shell','su_attempted','num_root','num_file_creations',
     'num_shells','num_access_files','num_outbound_cmds','is_host_login','is_guest_login','count','srv_count',
@@ -27,6 +20,15 @@ COLS = str(['duration','protocol_type','flag','src_bytes','dst_bytes','land','wr
     'srv_diff_host_rate','dst_host_count','dst_host_srv_count','dst_host_same_srv_rate','dst_host_diff_srv_rate',
     'dst_host_same_src_port_rate','dst_host_srv_diff_host_rate','dst_host_serror_rate','dst_host_srv_serror_rate',
     'dst_host_rerror_rate','dst_host_srv_rerror_rate','target'])
+MODEL_NAME = 'if'
+SAVE = True
+SAVE_PATH = './models/'
+
+# Isolation Forest hyperparameters
+CONTAMINATION = .1
+N_ESTIMATORS = 50
+MAX_SAMPLES = .8
+MAX_FEATURES = 1.
 
 def train(X,args):
     """ Fit Isolation Forest. """
@@ -36,7 +38,7 @@ def train(X,args):
     clf.fit(X)
     
     if args.save: # save model
-        with open(args.save_path + 'model.pickle', 'wb') as f:
+        with open(args.save_path + args.model_name + '.pickle', 'wb') as f:
             pickle.dump(clf,f)
 
 def run(args):
@@ -67,6 +69,7 @@ if __name__ == '__main__':
     parser.add_argument('--n_estimators',type=int,default=N_ESTIMATORS)
     parser.add_argument('--max_samples',type=float,default=MAX_SAMPLES)
     parser.add_argument('--max_features',type=float,default=MAX_FEATURES)
+    parser.add_argument('--model_name',type=str,default=MODEL_NAME)
     parser.add_argument('--save', default=SAVE, action='store_false')
     parser.add_argument('--save_path',type=str,default=SAVE_PATH)
     args = parser.parse_args()
