+++ components/outlier-detection/seq2seq-lstm/README.md
@@ -10,8 +10,15 @@ The implemented seq2seq outlier detector aims to predict anomalies in a sequence
 
 The architecture of the seq2seq model is defined in ```model.py``` and it is trained by running the ```train.py``` script. The ```OutlierSeq2SeqLSTM``` class loads a pre-trained model and makes predictions on new data.
 
-A detailed explanation of the implementation and usage of the seq2seq model as an outlier detector can be found in the [seq2seq_lstm_doc](./seq2seq_lstm_doc.ipynb) notebook.
+A detailed explanation of the implementation and usage of the seq2seq model as an outlier detector can be found in the [seq2seq documentation](./doc.md).
 
 ## Running on Seldon
 
 An end-to-end example running a seq2seq outlier detector on GCP or Minikube using Seldon to identify anomalies in ECGs is available [here](./seq2seq_lstm.ipynb).
+
+Docker images to use the generic Mahalanobis outlier detector as a model or transformer can be found on Docker Hub:
+* [seldonio/outlier-s2s-lstm-model](https://hub.docker.com/r/seldonio/outlier-s2s-lstm-model)
+* [seldonio/outlier-s2s-lstm-transformer](https://hub.docker.com/r/seldonio/outlier-s2s-lstm-transformer)
+
+A model docker image specific for the demo is also available:
+* [seldonio/outlier-s2s-lstm-model-demo](https://hub.docker.com/r/seldonio/outlier-s2s-lstm-model-demo)
