+++ components/outlier-detection/vae/OutlierVAE.py
@@ -1,50 +1,27 @@
 import numpy as np
-import pickle
-import random
 
-from model import model
+from CoreVAE import CoreVAE
 from utils import flatten, performance, outlier_stats
 
 
-class OutlierVAE(object):
+class OutlierVAE(CoreVAE):
     """ Outlier detection using variational autoencoders (VAE).
     
-    Arguments:
-        - threshold: (float): reconstruction error (mse) threshold used to classify outliers
-        - reservoir_size (int): number of observations kept in memory using reservoir sampling used for mean and stdev
+    Parameters
+    ----------
+        threshold (float) :  reconstruction error (mse) threshold used to classify outliers
+        reservoir_size (int) : number of observations kept in memory using reservoir sampling
      
-    Functions:
-        - reservoir_sampling: applies reservoir sampling to incoming data
-        - predict: detect and return outliers
-        - send_feedback: add target labels as part of the feedback loop
-        - metrics: return custom metrics
+    Functions
+    ----------
+        send_feedback : add target labels as part of the feedback loop
+        metrics : return custom metrics
     """
+    
     def __init__(self,threshold=10,reservoir_size=50000,model_name='vae',load_path='./models/'):
         
-        self.threshold = threshold
-        self.reservoir_size = reservoir_size
-        self.batch = []
-        self.N = 0 # total sample count up until now for reservoir sampling
-        
-        # load model architecture parameters
-        with open(load_path + model_name + '.pickle', 'rb') as f:
-            n_features, hidden_layers, latent_dim, hidden_dim, output_activation = pickle.load(f)
-            
-        # instantiate model
-        self.vae = model(n_features,hidden_layers=hidden_layers,latent_dim=latent_dim,
-                         hidden_dim=hidden_dim,output_activation=output_activation)
-        self.vae.load_weights(load_path + model_name + '_weights.h5') # load pretrained model weights
-        self.vae._make_predict_function()
-        
-        # load data preprocessing info
-        with open(load_path + 'preprocess_' + model_name + '.pickle', 'rb') as f:
-            preprocess = pickle.load(f)
-        self.preprocess, self.clip, self.axis = preprocess[:3]
-        if self.preprocess=='minmax':
-            self.xmin, self.xmax = preprocess[3:5]
-            self.min, self.max = preprocess[5:]
-        elif self.preprocess=='standardized':
-            self.mu, self.sigma = preprocess[3:]
+        super().__init__(threshold=threshold,reservoir_size=reservoir_size,
+                         model_name=model_name,load_path=load_path)
         
         self._predictions = []
         self._labels = []
@@ -52,83 +29,31 @@ class OutlierVAE(object):
         self.roll_window = 100
         self.metric = [float('nan') for i in range(18)]
         
-    
-    def reservoir_sampling(self,X,update_stand=False):
-        """ Keep batch of data in memory using reservoir sampling. """
-        for item in X:
-            self.N+=1
-            if len(self.batch) < self.reservoir_size:
-                self.batch.append(item)
-            else:
-                s = int(random.random() * self.N)
-                if s < self.reservoir_size:
-                    self.batch[s] = item
-        
-        if update_stand:
-            if self.preprocess=='minmax':
-                self.xmin = np.array(self.batch).min(axis=self.axis)
-                self.xmax = np.array(self.batch).max(axis=self.axis)
-            elif self.preprocess=='standardized':
-                self.mu = np.array(self.batch).mean(axis=self.axis)
-                self.sigma = np.array(self.batch).std(axis=self.axis)
-        return
-
-    
-    def predict(self,X,feature_names):
-        """ Detect outliers from mse using the threshold. 
-        
-        Arguments:
-            - X: input data
-            - feature_names
-        """
-        
-        # clip data per feature
-        X = np.clip(X,[-c for c in self.clip],self.clip)
-    
-        if self.N < self.reservoir_size:
-            update_stand = False
-        else:
-            update_stand = True
-            
-        self.reservoir_sampling(X,update_stand=update_stand)
-        
-        # apply scaling
-        if self.preprocess=='minmax':
-            X_scaled = ((X - self.xmin) / (self.xmax - self.xmin)) * (self.max - self.min) + self.min
-        elif self.preprocess=='standardized':
-            X_scaled = (X - self.mu) / (self.sigma + 1e-10)
-        
-        # sample latent variables and calculate reconstruction errors
-        N = 10
-        mse = np.zeros([X.shape[0],N])
-        for i in range(N):
-            preds = self.vae.predict(X_scaled)
-            mse[:,i] = np.mean(np.power(X_scaled - preds, 2), axis=1)
-        self.mse = np.mean(mse, axis=1)
-        self._mse.append(self.mse)
-        self._mse = flatten(self._mse)
-        
-        # make prediction
-        self.prediction = np.array([1 if e > self.threshold else 0 for e in self.mse]).astype(int)
-        self._predictions.append(self.prediction)
-        self._predictions = flatten(self._predictions)
-        
-        return self.prediction
-    
-    
+   
     def send_feedback(self,X,feature_names,reward,truth):
         """ Return outlier labels as part of the feedback loop.
         
-        Arguments:
-            - X: input data
-            - feature_names
-            - reward
-            - truth: outlier labels
+        Parameters
+        ----------
+            X : array of the features sent in the original predict request
+            feature_names : array of feature names. May be None if not available.
+            reward (float): the reward
+            truth : array with correct value (optional)
         """
+        _ = super().send_feedback(X,feature_names,reward,truth)
+        
+        # historical reconstruction errors and predictions
+        self._mse.append(self.mse)
+        self._mse = flatten(self._mse)
+        self._predictions.append(self.prediction)
+        self._predictions = flatten(self._predictions)
+        
+        # target labels
         self.label = truth
         self._labels.append(self.label)
         self._labels = flatten(self._labels)
         
+        # performance metrics
         scores = performance(self._labels,self._predictions,roll_window=self.roll_window)
         stats = outlier_stats(self._labels,self._predictions,roll_window=self.roll_window)
         
@@ -137,9 +62,9 @@ class OutlierVAE(object):
         for c in convert: # convert from np to native python type to jsonify
             metric.append(np.asscalar(np.asarray(c)))
         self.metric = metric
-            
-        return
-
+        
+        return []
+    
     
     def metrics(self):
         """ Return custom metrics.
@@ -154,8 +79,8 @@ class OutlierVAE(object):
             err = float('nan')
             y_true = float('nan')
         else:
-            pred = int(self._predictions[-2])
-            err = self._mse[-2]
+            pred = int(self._predictions[-1])
+            err = self._mse[-1]
             y_true = int(self.label[0])
                          
         is_outlier = {"type":"GAUGE","key":"is_outlier","value":pred}
