--- helm-charts/seldon-od-if/Chart.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-description: Seldon Core isolation forest outlier detection model template
-keywords:
-- kubernetes
-- machine-learning
-name: seldon-od-if
-sources:
-- https://github.com/SeldonIO/seldon-core
-version: 0.1
