--- helm-charts/seldon-od-if/README.md
@@ -1,4 +1,0 @@
-# Single Model with Outlier Detector
-
-This chart provides a Seldon Deployment with an outlier detector using Isolation Forests.
-
