--- helm-charts/seldon-od-md/values.yaml
@@ -1,15 +1,0 @@
-name: outlier-detector-md
-model:
-  image:
-    name: seldonio/outlier-mahalanobis:0.1
-  name: outlier-mahalanobis
-  threshold: 25
-  n_components: 3
-  n_stdev: 3
-  start_clip: 50
-  max_n: -1
-replicas: 1
-# Add oauth key and secret if using the default API Oauth Gateway for ingress
-oauth:
-  key: 
-  secret:
