+++ helm-charts/seldon-od-model/Chart.yaml
@@ -3,7 +3,7 @@ description: Seldon Core outlier detection model template
 keywords:
 - kubernetes
 - machine-learning
-name: seldon-od-vae
+name: seldon-od-model
 sources:
 - https://github.com/SeldonIO/seldon-core
 version: 0.1
