--- helm-charts/seldon-od-s2s-lstm/README.md
@@ -1,4 +1,0 @@
-# Single Model with Outlier Detector
-
-This chart provides a Seldon Deployment with an outlier detector using a Sequence-to-Sequence LSTM model.
-
