--- helm-charts/seldon-od-s2s-lstm/templates/model.json
@@ -1,64 +1,0 @@
-{
-    "apiVersion": "machinelearning.seldon.io/v1alpha2",
-    "kind": "SeldonDeployment",
-    "metadata": {
-        "labels": {
-            "app": "seldon"
-        },
-        "name": "{{ .Release.Name }}"
-    },
-    "spec": {
-        "name": "{{ .Release.Name }}",
-{{- if .Values.oauth.key }}        
-        "oauth_key": "{{ .Values.oauth.key }}",
-        "oauth_secret": "{{ .Values.oauth.secret }}",
-{{- end }}
-        "predictors": [
-            {
-                "componentSpecs": [{
-                    "spec": {
-                        "containers": [
-                            {
-                                "image": "{{ .Values.model.image.name }}",
-                                "imagePullPolicy": "IfNotPresent",
-                                "name": "{{ .Values.model.name }}",
-                                "resources": {
-                                    "requests": {
-                                        "memory": "1Mi"
-                                    }
-                                }
-                            }
-                        ],
-                        "terminationGracePeriodSeconds": 1
-                    }}			   
-	        ],
-                "graph":		 
-		{
-                    "children": [],
-                    "name": "{{ .Values.model.name }}",
-                    "endpoint": {
-			"type" : "REST"
-		    },
-                    "type": "MODEL",
-                    "parameters": [
-			{
-			    "name": "threshold",
-			    "value": "{{ .Values.model.threshold }}",
-			    "type": "FLOAT"
-			},
-			{
-			    "name": "reservoir_size",
-			    "value": "{{ .Values.model.reservoir_size }}",
-			    "type": "INT"
-			}
-		    ],
-                },		 			 
-                "name": "{{ .Release.Name }}",
-                "replicas": {{ .Values.replicas }},
-		"labels": {
-		    "version" : "v1"
-		}
-            }
-        ]
-    }
-}
