+++ helm-charts/seldon-od-transformer/Chart.yaml
@@ -1,9 +1,9 @@
 apiVersion: v1
-description: Seldon Core outlier detection model template
+description: Seldon Core outlier detection transformer template
 keywords:
 - kubernetes
 - machine-learning
-name: seldon-od-s2s-lstm
+name: seldon-od-transformer
 sources:
 - https://github.com/SeldonIO/seldon-core
 version: 0.1
