+++ helm-charts/seldon-od-transformer/README.md
@@ -1,0 +1,10 @@
+# Outlier Detector as Transformer
+
+This chart provides a Seldon Deployment with an outlier detector used as a transformer with a single model. 
+
+Available outlier detectors are:
+- [Sequence-to-Sequence LSTM](../../components/outlier-detection/seq2seq-lstm)
+- [Variational Auto-Encoder](../../components/outlier-detection/vae)
+- [Isolation Forest](../../components/outlier-detection/isolation-forest)
+- [Mahalanobis Distance](../../components/outlier-detection/mahalanobis)
+
