+++ helm-charts/seldon-od-transformer/values.yaml
@@ -1,0 +1,51 @@
+name: seldon-od-transformer
+model:
+  image:
+    name: seldonio/mock_classifier:1.0
+  name: classifier
+outlierDetection:
+  enabled: true
+  name: outlier-s2s
+  type: seq2seq
+  vae:
+    threshold: 10
+    reservoir_size: 50000
+    model_name: vae
+    load_path: ./models/
+    image:
+      name: seldonio/outlier-vae-tranformer:0.1
+  mahalanobis:
+    threshold: 25
+    n_components: 3
+    n_stdev: 3
+    start_clip: 50
+    max_n: -1
+    image:
+      name: seldonio/outlier-mahalanobis-tranformer:0.1
+  seq2seq:
+    threshold: 0.003
+    reservoir_size: 50000
+    model_name: seq2seq
+    load_path: ./models/
+    image:
+      name: seldonio/outlier-s2s-lstm-tranformer:0.1
+  isolationforest:
+    threshold: 0
+    model_name: if
+    load_path: ./models/
+    image:
+      name: seldonio/outlier-if-tranformer:0.1
+  parameterTypes:
+    threshold: FLOAT
+    reservoir_size: INT
+    model_name: STRING
+    load_path: STRING
+    n_components: INT
+    n_stdev: FLOAT
+    start_clip: INT
+    max_n: INT
+replicas: 1
+# Add oauth key and secret if using the default API Oauth Gateway for ingress
+oauth:
+  key: 
+  secret:
