--- helm-charts/seldon-od-vae/values.yaml
@@ -1,12 +1,0 @@
-name: outlier-detector-vae
-model:
-  image:
-    name: seldonio/outlier-vae:0.1
-  name: outlier-vae
-  threshold: 10
-  reservoir_size: 50000
-replicas: 1
-# Add oauth key and secret if using the default API Oauth Gateway for ingress
-oauth:
-  key: 
-  secret:
